Integrated spoonacular into main, unit testing, HTML changes
Spoonacular calls are displayed onto the actual web page, changed
some of the foods to prevent them from erroring. Unit tests are
failing but that is due to that fact that I have no idea why
the tests are refusing to interpret json correctly. Changed the
HTML so it fits in with corresponding code